Switched to using .ui file for MainWindow.
As the MainWindow contents were getting complex it was
time to switch to Qt .ui files. This commit adds a
main_window.ui file for the MainWindow as well as the
code generated by pyside-uic in ui_main_window.py .

The main window contains a tab widget for open circuit files
as well as a menu bar and, at least for now, replaces
history_window and history_dock_widget as their custom
behavior was so minimal it made more sense to have them
defined in the main_window.ui file and use promotions for
the sub-widgets that actually have custom behavior.