Add Settings class for loading/storing settings.
Added Settings class for handling LogikSim's configuration
needs. The class is meant to be used as a singleton retrievable
using the settings() function anywhere in the program. All
settings are generated as Qt properties with corresponding
change signals one can subscribe to.

There seem to be some strange binding issues which give me
trouble with working with doc-strings (encountered two, not
sure if they are connected). Worked around those for now.

Part of #8