Improve test execution speed by removing cargo culted delays.
As the vast majority of our GUI tests aren't relying on an event
loop there is really no need to wait for anything. I only added them
because I saw them in sample test code. Remove all delays not relying
on something happening asynchronously in an event loop and reduce the
delay for the ones that do. This should help our GUI tests scale in
the future as we add more of them.