Make settings decorator correctly set notify signal.
Previously made wrong assumptions about the use of the notify
signal in Qt properties. It is actually meant to expose the
signal we commit to emitting when our setter is called. Not
a way to tell Qt to emit this signal for us when the value
changes. Accordingly PYSIDE-261 was closed as invalid.