Fix sporadic modal dialog test failures by adding delays.
Unfortunately it seems like there can be differing delays
when opening and closing modal dialogs. On my local box this
might be caused by animations and other slowdowns which made
the issue reprodible locally for me by running the test-suite
in a loop. These additional delays hopefully also fix the
issue on the CI systems for good.