Extend and unify element interface.
For the elements to properly interact they have to
have a unified interface. The simple_element specific
attach is replaced by a generic connect method on the
element (as well as a disconnect method).

Also introduced a reset function which is meant to be
used on circuit initialization to propagate the right
states throughout the circuit. For that it basically
creates an edge with the current output state on each
output of each element.