Introduce compound elements.
The idea of compounds is to allow grouping of multiple
elements inside of one element. To allow this the
compound element acts like a normal element with inputs
and outputs to the outside but has input and output
indirection capabilities that can be used to wire up
the internal elements to those virtual input and output
posts.

Edge transitions are - currently - directly forwarded
to the corresponding connected internal element. Edges
on pins not internally connected are ignored.