Introduce total event ordering to prevent event racing issues.
Without total event ordering we cannot properly
handle things occurring at the same time at the moment.

We can probably solve this another way but with the current
architecture if we have two edges on different inputs on the
same element at the same time we can't apply them without
scheduling an independent output event for each edge. This
isn't a problem as long as our output events stay ordered
correctly. If they don't we might end up in an invalid state.