Add component concept.
This commit adds the concept of a ComponentType, ComponentInstance
as well as ComponentLibrary. A ComponentType contains the basic
information about something that can be placed in a LogikSim circuit.
The most important part of that is a GUID uniquely identifying this
component.

A ComponentInstance can be instantiated using the ComponentType and
and contains additional instance specific meta-data. It is also the
actual class instance participating in the simulation. A ComponentInstance
always contains a reference to its ComponentType. A meta-data lookup
using get_metadata_field always first checks whether the field is available
in the instance before falling back to the ComponentType.

The ComponentLibrary is used to keep track of the different
ComponentTypes in the application and can be used to enumerate and
instantiate them.

All in all this machinery is supposed to enable the GUI to talk about
components to the backend without needing any strong coupling. GUI
specific attributes can be saved in additional meta-data fields on
instance or type the core doesn't have to look at. It also enables user
created components by making the uniquely identifiable and defining
a common way of creating them for the GUI.