Start integrating backend with frontend.
Currently non-functional. Not using multiprocessing module for now either
as it makes things needlessly complicated. Using python thread instead.

Basically all this commit is for is to get a feel for the pain points
in integration and to get a better idea on how to change the current item
creation to make it mesh seamlessly with what the backend is doing.

Added basic logging (might break later with multiprocessing).