Fix or workaround unit test failures and hangs.
Not sure why but PySide objects don't seem to get properly
cleaned up. Probably some issue with the GC and Qt's object
management not interacting correctly. This wouldn't be a
big problem in the tests but results in threads not being
properly joined on exit making test execution hang or even
crash depending on what the tread is doing when it is
forcefully terminated.

Everything seems to be fine in the main application but I
wasn't able to make it work in the unit tests. For now
the workaround is to manually join the threads in the
unit tests tear down function. Still hoping to figure out
a real solution instead of having to introduce a separate
function for telling stuff it's supposed to have been
"destroyed" by now.

Additionally quite a few tests created a GridScene GUI widget
without making sure we had an initialized Qt gui application
object to work with. This only worked because other unit tests
that happened to run beforehand created one.