Standardize message framing and introduce clock field.
Messages now are sent using a function applying standardized
framing consisting of a message type (formerly action) and
the current simulation clock.

The backend now send an alive signal with every processing
call to the controller to notify the frontend that it is
still actively processing and also to allow the frontend
to sync its notion of the current state of the simulation
clock.

The status bar in the main window is now used to display
the current simulation clock. This isn't meant to be a
permanent construct but to temporarily help with debugging.

Also the simulation clock is now starting at 0 instead of -1
because as it now runs without events it would take on values
between -1 and 0 otherwise.

The unit tests were adjusted to cope with these changes.