Added delay to connections and updated Interconnect element.
Every connection now has delay that is set on connect.

Interconnect now is quite similar to SimpleElement. It is
an element with one input and a variable number of outputs
which replicates the input states to the outputs with the
usual connection delay. The number of available outputs grows
as needed.