Implemented Output-Connector Animation
Unconnected Connectors can now have different width. For
that line drawing logic had to be adapted to snap to right positions.
This has the advantage that connected connectors are visually different
to unconnected ones.

Next step is to merge the display logic with LineTree, as there is
a substancial amount in common.