Caching edits while editing, without backend notifications.
Movements are now again implemented as simple moves (no Drag&Drop)
Also introduced a new class ConnectableItem that abstracts
the bookkeeping for caching connections. Already used this
to re-implement many parts of the Connector.

Selection mode now uses the new editing mode. During movement of
items changes are only cached locally, while when switching to other modes
the changes are send to the backend.

Moving LineTrees is not supported yet, since their connections
are send to the backend and bring the GUI into an inconsistent state.