Add batch commands backend command exception handling.
The interface can now be asked for a batch context version
that collects commands and sends them when exited. The backend
will execute all commands in a batch in one go which prevents
other frontends from interfering with them. All replies will be
tagged with a batch-id matching the request-id of the batch.

Additionally the backend will now catch exceptions in commands,
log them and send a special error message. For this to work every
command now has a request-id that the error can be in-response-to.