Implemented central update_connection method in the scene.
I was able to remove the complex connection caches in ConnectableItem
and replace them with a central update_connection method in the
scene. It is called after the scene becomes active again and
handles all disconnects and connects in a very controlled manny.

The result is a robust, but still simple, design compared to the
older solutions.

However, anticipation of connections for ConnectorItems does not yet work.