Add instantiated_frontend_item function.
This function is meant to be used to re-register already created
frontend item instances with the registry even though they were
deleted from the backend. This is useful during undo/redo where
creating new instances for items with the same ID would invalidate
the old item references we are still using.