Modify controller to support better testing.
Introduced a flag to enforce re-raising of exceptions encountered
in the controller so they make it out of the process function instead
of being caught and serialized. This way unit tests that don't test
error paths can fail exactly at the spot where the error occurred.

Added the possibility to specify the queue type when creating the
controller. This way the backend and controller tests doesn't have to
deal with multiprocessing.Queue idiosyncrasies when there isn't
actually any parallelism going on (nor wanted) in the unit tests.