Add on-connect state propagation.
Elements now propagate their output port state
when being connected to an input. Currently this
is implemented by passing the state in the
connected function and immediately applying it.
The simulation then relies on the outside env
to ensure the element is appropriately clocked.

For this the controller connect function schedules
a clock for the element. As this mustn't be executed
instantly a _scheduling_epsilon value was introduced
that defines a small delay events by the user which
are supposed to happen "instantly" are delayed by.

This delay is now also applied to user requested edges
which resolves previously existing issues where clock
was called multiple times for the same simulation time
when the simulation was paused and the user kept
scheduling edges for "now".